+++ CHANGELOG.md
@@ -5,6 +5,7 @@ ChangeLog
 - Fix for listing clients
 - Fix for add bookmark response
 - Fix encoding site title
+- Fix get personal options
 
 
 1.5.0 (2020-03-24)